requirements.txt
flask==3.0.2
requests==2.31.0
transformers==4.38.2
langdetect==1.0.9
python-dotenv==1.0.1
gunicorn==21.2.0